+++ packages/fcl-core/src/current-user/exec-service/index.js
@@ -4,9 +4,21 @@ import {getServiceRegistry} from "./plugins"
 import {getChainId} from "../../utils"
 import {VERSION} from "../../VERSION"
 import {configLens} from "../../default-config"
-const execStrategy = async ({service, body, config, opts}) => {
+import {checkWalletConnectEnabled} from "./wc-check"
+
+const AbortController =
+  globalThis.AbortController || require("abort-controller")
+
+export const execStrategy = async ({
+  service,
+  body,
+  config,
+  abortSignal,
+  customRpc,
+  opts,
+}) => {
   const strategy = getServiceRegistry().getStrategy(service.method)
-  return strategy({service, body, config, opts})
+  return strategy({service, body, config, abortSignal, customRpc, opts})
 }
 
 export async function execService({
@@ -15,7 +27,12 @@ export async function execService({
   config = {},
   opts = {},
   platform,
+  abortSignal = new AbortController().signal,
+  execStrategy: _execStrategy,
 }) {
+  // Notify the developer if WalletConnect is not enabled
+  checkWalletConnectEnabled()
+
   msg.data = service.data
   const execConfig = {
     services: await configLens(/^service\./),
@@ -31,12 +48,14 @@ export async function execService({
   }
 
   try {
-    const res = await execStrategy({
+    const res = await (_execStrategy || execStrategy)({
       service,
       body: msg,
       config: execConfig,
       opts,
+      abortSignal,
     })
+
     if (res.status === "REDIRECT") {
       invariant(
         service.type === res.data.type,
@@ -47,6 +66,7 @@ export async function execService({
         msg,
         config: execConfig,
         opts,
+        abortSignal,
       })
     } else {
       return res