+++ packages/fcl-wc/src/index.ts
@@ -1,0 +1,4 @@
+export {SERVICE_PLUGIN_NAME, WC_SERVICE_METHOD} from "./service"
+export {init, initLazy, getSignClient} from "./fcl-wc"
+export {createSessionProposal, request} from "./session"
+export {FLOW_METHODS} from "./constants"