+++ packages/fcl-wc/src/utils.ts
@@ -1,18 +1,8 @@
 import {log, LEVELS} from "@onflow/util-logger"
 import {invariant} from "@onflow/util-invariant"
-import * as fcl from "@onflow/fcl"
+import * as fclCore from "@onflow/fcl-core"
 
-export let CONFIGURED_NETWORK = null
-
-export const setConfiguredNetwork = async () => {
-  CONFIGURED_NETWORK = await fcl.getChainId()
-  invariant(
-    CONFIGURED_NETWORK,
-    "FCL Configuration value for 'flow.network' is required"
-  )
-}
-
-const makeFlowServicesFromWallets = wallets => {
+const makeFlowServicesFromWallets = (wallets: any[]) => {
   return Object.values(wallets)
     .filter(w => w.app_type === "wallet")
     .map(wallet => {
@@ -37,10 +27,11 @@ const makeFlowServicesFromWallets = wallets => {
     })
 }
 
-export const fetchFlowWallets = async projectId => {
+export const fetchFlowWallets = async (projectId: string) => {
   try {
+    const network = await fclCore.getChainId()
     const wcApiWallets = await fetch(
-      `https://explorer-api.walletconnect.com/v3/wallets?projectId=${projectId}&chains=flow:${CONFIGURED_NETWORK}&entries=5&page=1`
+      `https://explorer-api.walletconnect.com/v3/wallets?projectId=${projectId}&chains=flow:${network}&entries=5&page=1`
     ).then(res => res.json())
 
     if (wcApiWallets?.count > 0) {
@@ -49,11 +40,13 @@ export const fetchFlowWallets = async projectId => {
 
     return []
   } catch (error) {
-    log({
-      title: `${error.name} Error fetching wallets from WalletConnect API`,
-      message: error.message,
-      level: LEVELS.error,
-    })
+    if (error instanceof Error) {
+      log({
+        title: `${error.name} Error fetching wallets from WalletConnect API`,
+        message: error.message,
+        level: LEVELS.error,
+      })
+    }
   }
 }