+++ packages/fcl/src/utils/web/render-frame.js
@@ -27,7 +27,7 @@ export function renderFrame(src) {
   const $frame = document.createElement("iframe")
   $frame.src = src
   $frame.id = FRAME
-  $frame.allow = "usb *; hid *"
+  $frame.allow = "usb *; hid *; clipboard-write"
   $frame.frameBorder = "0"
   $frame.style.cssText = FRAME_STYLES
   document.body.append($frame)